+++ webpack/app/lazysizes.js
@@ -1,0 +1,2 @@
+import 'lazysizes/plugins/twitter/ls.twitter.js';
+import 'lazysizes/lazysizes-umd.js';